+++ worker/src/Utils/Crypto.cpp
@@ -66,7 +66,7 @@ namespace Utils
 
 		// Init the vrypto seed with a random number taken from the address
 		// of the seed variable itself (which is random).
-		Crypto::seed = (uint32_t)(unsigned long)&Crypto::seed;
+		Crypto::seed = (uint32_t)(uintptr_t)&Crypto::seed;
 
 		// Create an OpenSSL HMAC_CTX context for HMAC SHA1 calculation.
 		HMAC_CTX_init(&Crypto::hmacSha1Ctx);
